% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGvizCode.R
\name{generateGvizCodeforSMF}
\alias{generateGvizCodeforSMF}
\title{Visualize Single Molecule Footprint Patterns with Ideogram and Gene Tracks}
\usage{
generateGvizCodeforSMF(
  organism = "demlanogaster",
  model = "S2",
  condition = "WT",
  genome_assembly = "dm6",
  type = "dSMF",
  chromosome = "chr2L",
  start = 480290,
  end = 480320,
  tr = "fp_and_mvec",
  label = "peak229",
  span_left = 150,
  span_right = 150,
  remove_dup = FALSE,
  fp_cap = 50,
  gviz_left = 1000,
  gviz_right = 1000,
  target_dir = ""
)
}
\arguments{
\item{organism}{Organism code (default: "dmelanogaster")}

\item{model}{Biological model/system (default: "S2" cells)}

\item{condition}{Experimental condition (default: "WT")}

\item{genome_assembly}{Genome version (default: "dm6")}

\item{type}{Data type ("dSMF" = dual enzyme Single Molecule Footprinting)}

\item{chromosome}{Chromosome ID (e.g., "chr2L")}

\item{start}{Genomic start position (numeric/character)}

\item{end}{Genomic end position (numeric/character)}

\item{tr}{Track name for BigBed file resource (default: "fp_and_mvec")}

\item{label}{Plot title annotation}

\item{span_left}{Upstream window size from region (default: 150)}

\item{span_right}{Downstream window size from region (default: 150)}

\item{remove_dup}{Remove duplicate reads? (default: FALSE)}

\item{fp_cap}{Maximum footprint value for y-axis scaling (default: 50)}

\item{gviz_left}{left zoom (bases) from the start}

\item{gviz_right}{right zoom (bases) from the end}

\item{target_dir}{destination directory (must be the same used in `plotFootprints` function)}
}
\value{
Saves a heatmap in pdf, png, and eps file formats.
}
\description{
Generates Gviz-compatible R code that puts SMF footprints below ideogram and genetracks. This call assumes that you have already ran `plotFootprints` function. 
Designed for Drosophila melanogaster analysis with expandable parameters for other organisms.
}
\details{
This function generates a R code running which one can get ideogram, genetracks and SMF heatmap altogether in one plot.
}
\examples{
# Basic usage with default parameters
generateGvizCodeforSMF(organism = "mmusculus", model = "16cell", 
                         condition = "WT", genome_assembly = "mm10",
                         type = "SMF", chromosome = "chr1",start = 191718250,
                         end = 191718280, tr = "16cell", label = "tss",
                         fp_cap = 50, remove_dup = FALSE, gviz_left = 500,
                         gviz_right = 500, target_dir = "")
                         

}
