% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plotMethylationCallsNanopore}
\alias{plotMethylationCallsNanopore}
\title{Visualize Single Molecule Footprint Patterns}
\usage{
plotMethylationCallsNanopore(
  organism = "scerevisiae",
  model = "BY4741 strain",
  condition = "WT",
  genome_assembly = "sacCer3",
  type = "SMF",
  chromosome = "chrIII",
  start = 114300,
  end = 114600,
  tr = "nanopore_meth_calls",
  label = "smac_seq",
  span_left = 1000,
  span_right = 1000,
  stride = 5,
  target_dir = ""
)
}
\arguments{
\item{organism}{Organism code (default: "scerevisiae")}

\item{model}{Biological model/system (default: "BY4741 strain")}

\item{condition}{Experimental condition (default: "WT")}

\item{genome_assembly}{Genome version (default: "sacCer3")}

\item{type}{Data type ("Nanopore" = dual enzyme Single Molecule Footprinting)}

\item{chromosome}{Chromosome ID (e.g., "chrIII")}

\item{start}{Genomic start position (numeric/character)}

\item{end}{Genomic end position (numeric/character)}

\item{tr}{Track name for BigBed file resource (default: "nanopore_meth_calls")}

\item{label}{Plot title annotation (default: "smac_seq")}

\item{span_left}{Upstream window size from region (default: 1000)}

\item{span_right}{Downstream window size from region (default: 1000)}

\item{stride}{sliding window used for singal aggregation}

\item{target_dir}{destination directory where outputs will be stored}
}
\value{
Saves a heatmap in pdf, png, and eps file formats. Also, occupancies are saved in a tsv file
}
\description{
Generates footprint plots from NOME-seq data within specified genomic regions.
Designed for Drosophila melanogaster analysis with expandable parameters for other organisms.
}
\details{
This function retrieves and visualizes DNA accessibility patterns from single-molecule data:
- Integrates with UCSC-style track hubs via `tr` parameter
- Automatically handles coordinate conversion for dm6 genome
- Implements dynamic y-axis scaling using `fp_cap`
}
\examples{
# Basic usage with default parameters
plotMethylationCallsNanopore()

# Custom genomic region analysis

plotMethylationCallsNanopore (
        organism = "scerevisiae", model = "BY4741 strain",
        condition = "WT", genome_assembly = "sacCer3",
        type = "SMF", chromosome = "chrIII",
        start = 114300, end = 114600,
        tr = "nanopore_meth_calls", label = "smac_seq",
        span_left = 1000, span_right = 1000, stride = 5, target_dir = "")
}
