% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{plotMethylationCallsNanoporeUsingLocalBigBed}
\alias{plotMethylationCallsNanoporeUsingLocalBigBed}
\title{Visualize aggregate methylation calls from ONT sequencing data}
\usage{
plotMethylationCallsNanoporeUsingLocalBigBed(
  bigBed = "",
  chromosome = "chrIII",
  start = 114300,
  end = 114600,
  label = "smac_seq",
  span_left = 1000,
  span_right = 1000,
  remove_dup = FALSE,
  stride = 5,
  target_dir = ""
)
}
\arguments{
\item{bigBed}{BigBed file path (e.g., "demo.bb")}

\item{chromosome}{Chromosome ID (e.g., "chr2L")}

\item{start}{Genomic start position (numeric/character)}

\item{end}{Genomic end position (numeric/character)}

\item{label}{Plot title annotation}

\item{span_left}{Upstream window size from region (default: 150)}

\item{span_right}{Downstream window size from region (default: 150)}

\item{remove_dup}{Remove duplicate reads? (default: FALSE)}

\item{stride}{a window size used for averaging the methylation call values}

\item{target_dir}{destination directory where outputs will be stored}
}
\value{
Saves a methylation calls (binary) heatmap in pdf, png, and eps file formats.
}
\description{
Generates heatmap of methylation calls from likes of SMAC-seq datasets.
currently has data from Yeast (sacCer3)
}
\details{
This function uses local SMF/dSMF bigBed file and plot heatmap
}
